Finished Model tests.

+++ eagleeye-msp-quote-service/src/main/java/com/shew/consulting/eagleeye/msp/quote/service/data/QuoteTestData.java
@@ -45,7 +45,7 @@ public class QuoteTestData {
                 }
             });
             quote.setSelections(selections);
-            quote.setCustomerId(i + 1);
+            quote.setCustomerId((long) (i + 1));
             quote.setTimestamp(Instant.now());
             quoteRepository.save(quote);
         });
